requirements.txt
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
aiohttp==3.9.1
beautifulsoup4==4.12.2
openai==1.3.7
tiktoken==0.5.1
tenacity==8.2.3
python-multipart==0.0.6
